Part 2
Extend the program functionality to allow the calculation of the final grade using either the average or the median. The output should display the chosen final grade calculation method. Additionally, modify the program to handle cases where the number of homework assignments (n) is unknown in advance, meaning the user decides when they have entered all the homework results. 
Implement this part using a std::vector container. For all next tasks, we will focus on the usage of vector.
Add to the program possibility to randomly generate homework and exam scores for students. 
Commit the updated result with the GitHub sub-repository - v0.1.